AITrainer/requirements.txt
ultralytics
google-generativeai
opencv-python
Pillow
python-dotenv
numpy

